pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "c4ai-scrpr"
version = "0.1.0"
description = "Event data platform powered by crawl4ai"
readme = "README.md"
requires-python = ">=3.11"
authors = [{name = "Codex", email = "codex@example.com"}]
license = {text = "MIT"}
dependencies = [
    "crawl4ai>=0.4.2",
    "httpx>=0.27.0",
    "aiohttp>=3.9.5",
    "uvloop>=0.19.0; sys_platform != 'win32'",
    "backoff>=2.2.1",
    "pydantic>=2.7.0",
    "pydantic-core>=2.18.0",
    "jsonschema>=4.22.0",
    "orjson>=3.10.0",
    "python-dotenv>=1.0.1",
    "structlog>=24.1.0",
    "python-rapidjson>=1.14",
    "PyYAML>=6.0.1",
    "beautifulsoup4>=4.12.3",
    "croniter>=2.0.5",
    "python-dateutil>=2.9.0.post0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.1",
    "pytest-asyncio>=0.23.7",
    "pytest-cov>=5.0.0",
    "hypothesis>=6.99.0",
    "ruff>=0.4.8",
    "black>=24.4.0",
    "mypy>=1.10.0",
    "types-aiofiles>=23.2.0.3",
    "types-python-dateutil>=2.9.0.20240316",
    "types-requests>=2.31.0.6"
]

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning"
]
addopts = "-q"

[tool.ruff]
line-length = 100
select = ["E", "F", "B", "I", "UP", "C4"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
warn_return_any = true
warn_unused_ignores = true
strict_optional = true
exclude = [
    "tests/fixtures",
]
